Refactor mortgage calculation command to improve validation
Updated the CLI command signature and logic to accept a single parameter type ("income" or "value") and its corresponding value. Added validation for invalid types and values with appropriate error messages. Adjusted tests, translations, and documentation to reflect these changes.

+++ tests/Unit/MortgageApiClientTest.php
@@ -54,7 +54,7 @@ class MortgageApiClientTest extends TestCase
         $this->expectException(\Exception::class);
         $this->expectExceptionMessage('Parameter person[0][income] is required.');
 
-        $this->mortgageClient->getMaximumMortgageByIncome(null);
+        $this->mortgageClient->getMaximumMortgageByIncome((int)null);
     }
 
     #[Test]
